server/requirements.txt
dnspython==2.7.0

# --- Core Frameworks ---
fastapi==0.115.12
uvicorn==0.34.0
starlette==0.46.1

# --- GraphQL ---
strawberry-graphql==0.263.0

# --- MongoDB + ODM ---
motor                    
beanie==1.18.0                   
pymongo==4.11.3          

# --- Pydantic + Types ---
pydantic[email]==1.10.12
pydantic_core==2.33.0
annotated-types==0.7.0
typing-inspection==0.4.0
typing_extensions==4.13.0

# --- JWT & Security ---
fastapi-jwt-auth
pyjwt[crypto]            
cryptography             
passlib[bcrypt]==1.7.4
bcrypt==3.2.0         
cffi                     
python-dotenv==1.1.0

# --- Utilities ---
httpx                    
python-dateutil==2.9.0.post0
packaging==24.2
colorama==0.4.6
click==8.1.8
six==1.17.0
idna==3.10
h11==0.14.0
anyio==4.9.0
sniffio==1.3.1

# --- Testing ---
pytest
